Added the spotbugs plugin and fixed potential issues as it suggested; and some cleanup.

+++ src/main/java/cpath/service/BiopaxConverter.java
@@ -137,10 +137,12 @@ public class BiopaxConverter {
       dataResponse.setProviders(providers(m));
       return dataResponse;
     } catch (Exception e) {
-      try {
-        Files.delete(tmpPath);
-      } catch (Exception ex) {
-        log.error(e.toString());
+      if(tmpPath != null) {
+        try {
+          Files.delete(tmpPath);
+        } catch (Exception ex) {
+          log.error(e.toString());
+        }
       }
       return new ErrorResponse(INTERNAL_ERROR, e);
     }
